+++ localization_common/include/localization_common/cloud_registration/ndt_omp_registration.hpp
@@ -30,8 +30,8 @@ public:
   NdtOmpRegistration(float res, float step_size, float trans_eps, int max_iter);
 
   bool set_target(const PointCloudPtr & target) override;
-  bool match(const PointCloudPtr & input, const Eigen::Matrix4f & initial_pose) override;
-  Eigen::Matrix4f get_final_pose() override;
+  bool match(const PointCloudPtr & input, const Eigen::Matrix4d & initial_pose) override;
+  Eigen::Matrix4d get_final_pose() override;
   double get_fitness_score() override;
   void print_info() override;
